Правки логики событий  хедера. улучшение разделения игровых экранов.
