+++ js/utils.js
@@ -141,4 +141,66 @@ const showGameScreen = (data) => {
   }
 };
 
-export {getElementFromTemplate, getScore, createTimer, getFastAnswersValue, getSlowAnswersValue, getCompletedLevelsValue, getLevelProgressBar, getStatsTitle, getStatsResult, createUserdata, userState, showGameScreen};
+const handleResultOfOneCardLevel = (data, firstCardInputsValue) => {
+  let answerIsSolved;
+  if (firstCardInputsValue === data.cards[0].rightAnswer) {
+    answerIsSolved = true;
+  } else {
+    userState.lives = userState.lives - 1;
+    answerIsSolved = false;
+  }
+  let answerOnCard = {
+    time: AnswerTime.NORMAL,
+    solved: answerIsSolved
+  };
+  userState.answers.push(answerOnCard);
+  return answerOnCard;
+};
+
+const handleResultOfTwoCardsLevel = (currentLevel, firstCardInputsValue, secondCardInputsValue) => {
+  let answerIsSolved;
+  if (firstCardInputsValue === currentLevel.cards[0].rightAnswer && secondCardInputsValue === currentLevel.cards[1].rightAnswer) {
+    answerIsSolved = true;
+  } else {
+    userState.lives = userState.lives - 1;
+    answerIsSolved = false;
+  }
+  let answerOnCard = {
+    time: AnswerTime.NORMAL,
+    solved: answerIsSolved
+  };
+  userState.answers.push(answerOnCard);
+  return answerOnCard;
+};
+
+const findDifferentCard = (firstCard, secondCard, thirdCard) => {
+  const firstCardEl = document.querySelector(`.game__option:first-child`);
+  const secondCardEl = document.querySelector(`.game__option:nth-child(2)`);
+  const thirdCardEl = document.querySelector(`.game__option:nth-child(3)`);
+  if (firstCard !== secondCard && firstCard !== thirdCard) {
+    return firstCardEl;
+  } else if (secondCard !== firstCard && secondCard !== thirdCard) {
+    return secondCardEl;
+  } else if (thirdCard !== firstCard && thirdCard !== firstCard) {
+    return thirdCardEl;
+  }
+  throw new Error(`Incorrect type of parameters. (need paint or photo string)`);
+};
+
+const handleResultOfThreeCardsLevel = (evt, firstCardAnswer, secondCardAnswer, thirdCardAnswer) => {
+  let answerIsSolved;
+  if (evt.target === findDifferentCard(firstCardAnswer, secondCardAnswer, thirdCardAnswer)) {
+    answerIsSolved = true;
+  } else {
+    userState.lives = userState.lives - 1;
+    answerIsSolved = false;
+  }
+  let answerOnCard = {
+    time: AnswerTime.NORMAL,
+    solved: answerIsSolved
+  };
+  userState.answers.push(answerOnCard);
+  return answerOnCard;
+};
+
+export {getElementFromTemplate, getScore, createTimer, getFastAnswersValue, getSlowAnswersValue, getCompletedLevelsValue, getLevelProgressBar, getStatsTitle, getStatsResult, createUserdata, userState, showGameScreen, handleResultOfTwoCardsLevel, handleResultOfOneCardLevel, handleResultOfThreeCardsLevel, findDifferentCard};
